[feature/home-more]: [feature/home-more] 홈에서 더 보기 화면 구현

+++ common/src/commonMain/kotlin/common/model/reseponse/paging/Pageable.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Pageable(
-    val offset: Int,
+    val offset: Long,
     val pageNumber: Int,
     val pageSize: Int,
     val paged: Boolean,
